+++ src/main/java/fi/dy/masa/malilib/compat/forge/ForgePlatformCompat.java
@@ -1,28 +1,20 @@
 package fi.dy.masa.malilib.compat.forge;
 
 import fi.dy.masa.malilib.compat.forge.register.ModConfigScreenRegister;
-import fi.dy.masa.malilib.compat.forge.register.impl.ModConfigScreenRegisterImpl;
-import net.minecraftforge.fml.ExtensionPoint;
-import net.minecraftforge.fml.ModLoadingContext;
-import net.minecraftforge.fml.network.FMLNetworkConstants;
-import org.apache.commons.lang3.tuple.Pair;
 
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
+@Deprecated
 public class ForgePlatformCompat {
     private static final ThreadLocal<ForgePlatformCompat> forgePlatform = ThreadLocal.withInitial(ForgePlatformCompat::new);
-    private static final Map<String, ModConfigScreenRegister> mods = new ConcurrentHashMap<>();
 
     public static ForgePlatformCompat getInstance() {
         return forgePlatform.get();
     }
 
     public ModConfigScreenRegister getMod(String id) {
-        return mods.computeIfAbsent(id, ModConfigScreenRegisterImpl::new);
+        return ForgePlatformUtils.getInstance().getMod(id);
     }
 
     public void getModClientExtensionPoint() {
-        ModLoadingContext.get().registerExtensionPoint(ExtensionPoint.DISPLAYTEST, () -> Pair.of(() -> FMLNetworkConstants.IGNORESERVERONLY, (a, b) -> true));
+        ForgePlatformUtils.getInstance().getClientModIgnoredServerOnly();
     }
 }