+++ CHANGELOG.md
@@ -1,2 +1,5 @@
 ## Changelog
 - fix icon missing issue
+- Deprecated `ForgePlatformCompat`, all function move to `ForgePlatformUtils`
+- Add old MaFgLib modid compat
+- Update Architectury Loom to 1.4